==> project/3-reglas_asociacion.py <==
# Empezar a tirar las predicciones


## Proceso supervisado

### Ahora sí balencear las variables para procesos supervisados


## Proceso no supervisado

### Ahora se deben balancear las variables


#> Resultado de cada proceso: Dataset smoteado
#> Entender lógica y visualmente qué predicciones aspectos se pueden conocer de la información
